Prepare cmake for header-source split; perform split for calculation-related functions
