+++ include/kettle/calculations/measurements.hpp
@@ -1,117 +1,25 @@
 #pragma once
 
-#include <algorithm>
 #include <cstddef>
-#include <optional>
-#include <random>
-#include <stdexcept>
-#include <string>
 #include <map>
+#include <optional>
+#include <string>
 #include <vector>
 
 #include "kettle/calculations/probabilities.hpp"
 #include "kettle/circuit/circuit.hpp"
-#include "kettle/common/mathtools.hpp"
-#include "kettle/common/prng.hpp"
-#include "kettle/simulation/simulate.hpp"
 #include "kettle/state/state.hpp"
-#include "kettle/state/marginal.hpp"
 
 /*
     This file contains code components to perform measurements of the state.
 */
 
-namespace impl_ket
-{
-
-class ProbabilitySampler_
-{
-public:
-    explicit ProbabilitySampler_(const std::vector<double>& probabilities, std::optional<int> seed = std::nullopt)
-        : cumulative_ {calculate_cumulative_sum(probabilities)}
-        , prng_ {get_prng_(seed)}
-    {
-        const auto max_prob = cumulative_.back();
-        const auto offset = impl_ket::cumulative_end_offset(cumulative_);
-        uniform_dist_ = std::uniform_real_distribution<double> {0.0, max_prob - offset};
-    }
-
-    auto operator()() -> std::size_t
-    {
-        const auto prob = uniform_dist_(prng_);
-
-        const auto it_state = std::ranges::lower_bound(cumulative_, prob);
-
-        if (it_state == cumulative_.end()) {
-            throw std::runtime_error {
-                "LOGIC BUG: Ended up with measurement of state past end of cumulative\n"
-                "probability distribution, which shouldn't happen?"};
-        }
-
-        const auto i_state = static_cast<std::size_t>(std::distance(cumulative_.begin(), it_state));
-
-        return i_state;
-    }
-
-private:
-    std::vector<double> cumulative_;
-    std::mt19937 prng_;
-    std::uniform_real_distribution<double> uniform_dist_;
-};
-
-inline auto build_marginal_bitmask_(
-    const std::vector<std::size_t>& marginal_qubits,
-    std::size_t n_qubits
-) -> std::vector<std::uint8_t>
-{
-    const auto is_in_range = [&](auto i) { return i >= n_qubits; };
-    if (std::ranges::any_of(marginal_qubits, is_in_range))
-    {
-        throw std::runtime_error {"ERROR: marginal qubit index out of range."};
-    }
-
-    auto marginal_bitmask = std::vector<std::uint8_t>(n_qubits, 0);
-    for (auto index : marginal_qubits) {
-        marginal_bitmask[index] = 1;
-    }
-
-    return marginal_bitmask;
-}
-
-}  // namespace impl_ket
-
 namespace ket
 {
 
-inline auto memory_to_counts(const std::vector<std::size_t>& measurements)
-    -> std::map<std::size_t, std::size_t>
-{
-    auto map = std::map<std::size_t, std::size_t> {};
+auto memory_to_counts(const std::vector<std::size_t>& measurements) -> std::map<std::size_t, std::size_t>;
 
-    // REMINDER: if the entry does not exist, `std::map` will first initialize it to 0
-    for (auto i_state : measurements) {
-        ++map[i_state];
-    }
-
-    return map;
-}
-
-inline auto memory_to_fractions(const std::vector<std::size_t>& measurements) -> std::map<std::size_t, double>
-{
-    auto map = std::map<std::size_t, double> {};
-
-    // REMINDER: if the entry does not exist, `std::map` will first initialize it to 0
-    for (auto i_state : measurements) {
-        map[i_state] += 1.0;
-    }
-
-    const auto n_measurements = static_cast<double>(measurements.size());
-    for (auto& pair : map) {
-        map[pair.first] /= n_measurements;
-    }
-
-    return map;
-}
+auto memory_to_fractions(const std::vector<std::size_t>& measurements) -> std::map<std::size_t, double>;
 
 /*
     Performs measurements of the QuantumState using its probabilities. The measurements
@@ -131,149 +39,61 @@ inline auto memory_to_fractions(const std::vector<std::size_t>& measurements) ->
       - memory complexity: O(max(2^n, k))
       - time complexity: O(k * 2^n)
 */
-inline auto perform_measurements_as_memory(
+auto perform_measurements_as_memory(
     const std::vector<double>& probabilities_raw,
     std::size_t n_shots,
     std::optional<int> seed = std::nullopt
-) -> std::vector<std::size_t>
-{
-    auto sampler = impl_ket::ProbabilitySampler_ {probabilities_raw, seed};
+) -> std::vector<std::size_t>;
 
-    auto measurements = std::vector<std::size_t> {};
-    measurements.reserve(n_shots);
-
-    for (std::size_t i_shot {0}; i_shot < n_shots; ++i_shot) {
-        const auto i_state = sampler();
-        measurements.push_back(i_state);
-    }
-
-    return measurements;
-}
-
-inline auto perform_measurements_as_memory(
+auto perform_measurements_as_memory(
     const QuantumState& state,
     std::size_t n_shots,
     const QuantumNoise* noise = nullptr,
     std::optional<int> seed = std::nullopt
-) -> std::vector<std::size_t>
-{
-    const auto probabilities_raw = calculate_probabilities_raw(state, noise);
-    return perform_measurements_as_memory(probabilities_raw, n_shots, seed);
-}
+) -> std::vector<std::size_t>;
 
-inline auto perform_measurements_as_counts_raw(
+auto perform_measurements_as_counts_raw(
     const std::vector<double>& probabilities_raw,
     std::size_t n_shots,
     std::optional<int> seed = std::nullopt
-) -> std::map<std::size_t, std::size_t>
-{
-    auto sampler = impl_ket::ProbabilitySampler_ {probabilities_raw, seed};
-    auto measurements = std::map<std::size_t, std::size_t> {};
+) -> std::map<std::size_t, std::size_t>;
 
-    // REMINDER: if the entry does not exist, `std::map` will first initialize it to 0
-    for (std::size_t i_shot {0}; i_shot < n_shots; ++i_shot) {
-        const auto i_state = sampler();
-        ++measurements[i_state];
-    }
-
-    return measurements;
-}
-
-inline auto perform_measurements_as_counts_raw(
+auto perform_measurements_as_counts_raw(
     const QuantumState& state,
     std::size_t n_shots,
     const QuantumNoise* noise = nullptr,
     std::optional<int> seed = std::nullopt
-) -> std::map<std::size_t, std::size_t>
-{
-    const auto probabilities_raw = calculate_probabilities_raw(state, noise);
-    return perform_measurements_as_counts_raw(probabilities_raw, n_shots, seed);
-}
+) -> std::map<std::size_t, std::size_t>;
 
-inline auto perform_measurements_as_counts_marginal(
+auto perform_measurements_as_counts_marginal(
     const std::vector<double>& probabilities_raw,
     std::size_t n_shots,
     const std::vector<std::size_t>& marginal_qubits = {},
     std::optional<int> seed = std::nullopt
-) -> std::map<std::string, std::size_t>
-{
-    if (!impl_ket::is_power_of_2(probabilities_raw.size())) {
-        throw std::runtime_error {"The number of probabilities must be a power of 2.\n"};
-    }
+) -> std::map<std::string, std::size_t>;
 
-    const auto n_qubits = impl_ket::log_2_int(probabilities_raw.size());
-    const auto marginal_bitmask = impl_ket::build_marginal_bitmask_(marginal_qubits, n_qubits);
-
-    auto sampler = impl_ket::ProbabilitySampler_ {probabilities_raw, seed};
-    auto measurements = std::map<std::string, std::size_t> {};
-
-    // the internal layout of the quantum state is little endian, so the probabilities are as well
-    const auto endian = ket::QuantumStateEndian::LITTLE;
-
-    // REMINDER: if the entry does not exist, `std::map` will first initialize it to 0
-    for (std::size_t i_shot {0}; i_shot < n_shots; ++i_shot) {
-        const auto i_state = sampler();
-        const auto bitstring = impl_ket::state_index_to_bitstring_marginal_(i_state, marginal_bitmask, endian);
-        ++measurements[bitstring];
-    }
-
-    return measurements;
-}
-
-inline auto perform_measurements_as_counts_marginal(
+auto perform_measurements_as_counts_marginal(
     const QuantumState& state,
     std::size_t n_shots,
     const std::vector<std::size_t>& marginal_qubits = {},
     const QuantumNoise* noise = nullptr,
     std::optional<int> seed = std::nullopt
-) -> std::map<std::string, std::size_t>
-{
-    const auto probabilities_raw = calculate_probabilities_raw(state, noise);
-    return perform_measurements_as_counts_marginal(probabilities_raw, n_shots, marginal_qubits, seed);
-}
+) -> std::map<std::string, std::size_t>;
 
-inline auto perform_measurements_as_counts_marginal(
+auto perform_measurements_as_counts_marginal(
     const QuantumCircuit& circuit,
     const QuantumState& original_state,
     std::size_t n_shots,
     const std::vector<std::size_t>& marginal_qubits = {},
     const QuantumNoise* noise = nullptr,
     std::optional<int> seed = std::nullopt
-) -> std::map<std::string, std::size_t>
-{
-    const auto n_qubits = circuit.n_qubits();
-    const auto marginal_bitmask = impl_ket::build_marginal_bitmask_(marginal_qubits, n_qubits);
+) -> std::map<std::string, std::size_t>;
 
-    // the internal layout of the quantum state is little endian, so the probabilities are as well
-    const auto endian = ket::QuantumStateEndian::LITTLE;
-
-    auto measurements = std::map<std::string, std::size_t> {};
-
-    for (std::size_t i {0}; i < n_shots; ++i) {
-        auto state = original_state;
-        ket::simulate(circuit, state);
-
-        const auto probabilities_raw = calculate_probabilities_raw(state, noise);
-        auto sampler = impl_ket::ProbabilitySampler_ {probabilities_raw, seed};
-
-        const auto i_state = sampler();
-        const auto bitstring = impl_ket::state_index_to_bitstring_marginal_(i_state, marginal_bitmask, endian);
-        ++measurements[bitstring];
-    }
-
-    return measurements;
-}
-
-inline auto perform_measurements_as_counts(
+auto perform_measurements_as_counts(
     const QuantumState& state,
     std::size_t n_shots,
     const QuantumNoise* noise = nullptr,
     std::optional<int> seed = std::nullopt
-) -> std::map<std::string, std::size_t>
-{
-    const auto probabilities_raw = calculate_probabilities_raw(state, noise);
-    const auto marginal_qubits = std::vector<std::size_t> {};
-    return perform_measurements_as_counts_marginal(probabilities_raw, n_shots, marginal_qubits, seed);
-}
+) -> std::map<std::string, std::size_t>;
 
 }  // namespace ket
